# Optimize DOM APIs