--- a/packages/react-native/Libraries/ReactNative/ReactFabricPublicInstance/__tests__/ReactFabricPublicInstance-itest.js
+++ b/packages/react-native/Libraries/ReactNative/ReactFabricPublicInstance/__tests__/ReactFabricPublicInstance-itest.js
@@ -14,7 +14,6 @@ import '@react-native/fantom/src/setUpDefaultReactNativeEnvironment';
 import type {HostInstance} from 'react-native';
 
 import ReactNativeElement from '../../../../src/private/webapis/dom/nodes/ReactNativeElement';
-import {getRawNativeDOMForTests} from '../../../../src/private/webapis/dom/nodes/specs/NativeDOM';
 import TextInputState from '../../../Components/TextInput/TextInputState';
 import View from '../../../Components/View/View';
 import ReactFabricHostComponent from '../ReactFabricHostComponent';
@@ -349,58 +348,5 @@ describe('ReactFabricPublicInstance', () => {
           .toJSX(),
       ).toEqual(<rn-view testID={'second test id'} />);
     });
-
-    // TODO: delete when NativeDOM.setNativeProps is NOT nullable.
-    // This logic is to ensure compatibility with old app versions without the native module method.
-    if (ReactNativeFeatureFlags.enableAccessToHostTreeInFabric()) {
-      let RawNativeDOM;
-      let originalSetNativeProps;
-
-      beforeAll(() => {
-        RawNativeDOM = nullthrows(getRawNativeDOMForTests());
-        originalSetNativeProps = RawNativeDOM.setNativeProps;
-      });
-
-      beforeEach(() => {
-        // $FlowExpectedError[cannot-write]
-        RawNativeDOM.setNativeProps = originalSetNativeProps;
-      });
-
-      it('should propagate changes to the host component (when NativeDOM.setNativeProps is not available)', () => {
-        // $FlowExpectedError[cannot-write]
-        RawNativeDOM.setNativeProps = null;
-
-        expect(RawNativeDOM.setNativeProps).toBeNull();
-
-        const root = Fantom.createRoot();
-        const nodeRef = createRef<HostInstance>();
-
-        Fantom.runTask(() => {
-          root.render(<View ref={nodeRef} testID="first test id" />);
-        });
-
-        expect(
-          root
-            .getRenderedOutput({
-              props: ['testID'],
-            })
-            .toJSX(),
-        ).toEqual(<rn-view testID={'first test id'} />);
-
-        const element = nullthrows(nodeRef.current);
-
-        Fantom.runTask(() => {
-          element.setNativeProps({testID: 'second test id'});
-        });
-
-        expect(
-          root
-            .getRenderedOutput({
-              props: ['testID'],
-            })
-            .toJSX(),
-        ).toEqual(<rn-view testID={'second test id'} />);
-      });
-    }
   });
 });