… packages/react-native/src/private/webapis/dom/oldstylecollections/NodeList.js
… packages/react-native/src/private/webapis/performance/Performance.js